// back-end/src/types/FormTypes.model.ts
export interface IFormSchema {
  firstName: string;
  surname: string;
  email: string;
  phoneNumber: number;
  gender: string;
  dob: string;
  comments: string;
}

export interface ISQL3FormSchema {
  id: string;
  firstName: string;
  surname: string;
  email: string;
  phoneNumber: number;
  gender: string;
  dob: string;
  comments: string;
}

export interface IParams {
  id: string;
}